fix: TS SDK updated and tested with new tests for payment x402 interceptor
